Add tests for backstage passes, failing tests are commented
The failing tests (and code review) reveals that quality is not being incremented correctly when a quality boundar is passed. The previous boundaries rules continue to apply. The SellIn date dependant quality rules should be applied _after_ the SellIn has been reduced, based ont he behaviour of the other items.